requirements.txt
hydra-core
rich
pyarrow
vaex
numpy==1.26.4
torch
tqdm
scikit-learn
transformers
torchmetrics
gensim
wandb
seaborn
wget
statsmodels
nest_asyncio
fastapi
uvicorn
streamlit
adjustText
psutil